fix(deploy-web): moved back home page to pages router
