+++ deploy-web/src/components/home/WelcomePanel.tsx
@@ -1,11 +1,13 @@
-import { Card, CardContent, CardHeader, Collapse, IconButton, List, ListItem, ListItemAvatar, ListItemText, useTheme } from "@mui/material";
+"use client";
 import { ReactNode, useState } from "react";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import RocketLaunchIcon from "@mui/icons-material/RocketLaunch";
-import CategoryIcon from "@mui/icons-material/Category";
-import SchoolIcon from "@mui/icons-material/School";
 import Link from "next/link";
 import { UrlService } from "@src/utils/urlUtils";
+import { Card, CardContent, CardHeader, CardTitle } from "@src/components/ui/card";
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from "@src/components/ui/collapsible";
+import { Rocket, SearchEngine, Learning, NavArrowDown } from "iconoir-react";
+import { Button } from "@src/components/ui/button";
+import { Avatar, AvatarFallback } from "@src/components/ui/avatar";
+import { cn } from "@src/utils/styleUtils";
 
 type Props = {
   children?: ReactNode;
@@ -13,58 +15,67 @@ type Props = {
 
 export const WelcomePanel: React.FC<Props> = () => {
   const [expanded, setExpanded] = useState(true);
-  const theme = useTheme();
 
   return (
-    <Card elevation={1}>
-      <CardHeader
-        title="Welcome to Cloudmos!"
-        titleTypographyProps={{ variant: "h3", sx: { fontSize: "1.25rem", fontWeight: "bold" } }}
-        sx={{ borderBottom: expanded ? `1px solid ${theme.palette.mode === "dark" ? theme.palette.grey[800] : theme.palette.grey[200]}` : "" }}
-        action={
-          <IconButton onClick={() => setExpanded(prev => !prev)}>
-            <ExpandMoreIcon />
-          </IconButton>
-        }
-      />
+    <Collapsible open={expanded} onOpenChange={setExpanded}>
+      <Card>
+        <CardHeader className="flex flex-row items-center justify-between">
+          <CardTitle className="text-2xl font-bold">Welcome to Akash Console!</CardTitle>
 
-      <Collapse in={expanded} timeout="auto" unmountOnExit>
-        <CardContent sx={{ padding: "0 !important" }}>
-          <List>
-            <ListItem>
-              <ListItemAvatar>
-                <RocketLaunchIcon fontSize="large" />
-              </ListItemAvatar>
-              <ListItemText
-                primary={<Link href={UrlService.getStarted()}>Getting started with Cloudmos</Link>}
-                secondary="Learn how to deploy your first docker container on Akash in a few click using Cloudmos."
-              />
-            </ListItem>
-            <ListItem>
-              <ListItemAvatar>
-                <CategoryIcon fontSize="large" />
-              </ListItemAvatar>
-              <ListItemText
-                primary={<Link href={UrlService.templates()}>Explore the marketplace</Link>}
-                secondary="Browse through the marketplace of pre-made solutions with categories like blogs, blockchain nodes and more!"
-              />
-            </ListItem>
-            <ListItem>
-              <ListItemAvatar>
-                <SchoolIcon fontSize="large" />
-              </ListItemAvatar>
-              <ListItemText
-                primary={
-                  <a target="_blank" rel="noopener noreferrer" href="https://docs.akash.network/">
+          <CollapsibleTrigger asChild>
+            <Button size="icon" variant="ghost" className="!m-0 rounded-full" onClick={() => setExpanded(prev => !prev)}>
+              <NavArrowDown fontSize="1rem" className={cn("transition-all duration-100", { ["rotate-180"]: expanded })} />
+            </Button>
+          </CollapsibleTrigger>
+        </CardHeader>
+
+        <CollapsibleContent>
+          <CardContent>
+            <ul className="space-y-4">
+              <li className="flex items-center">
+                <Avatar className="h-12 w-12">
+                  <AvatarFallback>
+                    <Rocket />
+                  </AvatarFallback>
+                </Avatar>
+
+                <div className="ml-4 flex flex-col">
+                  <Link href={UrlService.getStarted()}>Getting started with Akash Console</Link>
+                  <span>Learn how to deploy your first docker container on Akash in a few click using Console.</span>
+                </div>
+              </li>
+
+              <li className="flex items-center">
+                <Avatar className="h-12 w-12">
+                  <AvatarFallback>
+                    <SearchEngine />
+                  </AvatarFallback>
+                </Avatar>
+
+                <div className="ml-4 flex flex-col">
+                  <Link href={UrlService.templates()}>Explore the marketplace</Link>
+                  <span>Browse through the marketplace of pre-made solutions with categories like blogs, blockchain nodes and more!</span>
+                </div>
+              </li>
+
+              <li className="flex items-center">
+                <Avatar className="h-12 w-12">
+                  <AvatarFallback>
+                    <Learning />
+                  </AvatarFallback>
+                </Avatar>
+
+                <div className="ml-4 flex flex-col">
+                  <Link href="https://docs.akash.network/" target="_blank">
                     Learn more about Akash
-                  </a>
-                }
-                secondary="Want to know about the advantages of using a decentralized cloud compute marketplace?"
-              />
-            </ListItem>
-          </List>
-        </CardContent>
-      </Collapse>
-    </Card>
+                  </Link>
+                  <span>Want to know about the advantages of using a decentralized cloud compute marketplace?</span>
+                </div>
+              </li>
+            </ul>
+          </CardContent>
+        </CollapsibleContent>
+      </Card>
+    </Collapsible>
   );
 };
