+++ deploy-web/src/types/network.ts
@@ -7,6 +7,6 @@ export type Network = {
   chainRegistryName: string;
   versionUrl: string;
   rpcEndpoint?: string;
-  version: string;
+  version: string | null;
   enabled: boolean;
 };
